More isStateChange cleanup
We always want to update the state for whatever the command was
specifically for, but everything else can determine changes
automatically.

For example, if we send off() we want the status to confirm that the
device is off even if it already was, but it only needs to show that
effect is none if it was previously colorloop.