Adjust setLevel
Hue does not allow setting brightness to 0. Certain use-cases desire
setting brightness as low as possible, but keeping the bulb status to
on.

Also, the previous algorithm for setting brightness allowed for a
minimum brightness level of three. This introduces a new algorithm that
better converts the ST scale to Hue scale and properly allows setting
brightness level to 1.